pyproject.toml
[tool.poetry]
name = "os-computer-use"
version = "0.1.0"
description = ""
authors = ["James Murdza <james@jamesmurdza.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
e2b = "^1.0.3"
ipython = "^8.29.0"
python-dotenv = "^1.0.1"
gradio-client = "^1.5.0"
pillow = "^11.0.0"
e2b-desktop = "^1.0.2"
openai = "^1.59.5"
anthropic = "^0.44.0"
litellm = "^1.60.6"

[tool.poetry.scripts]
start = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
